Refactor.

+++ model/search_input.go
@@ -3,5 +3,5 @@ package model
 type WordsSearchInput struct {
 	Letters         string `json:"letters" binding:"required"`
 	AuxiliaryLetter string `json:"auxiliaryLetter"`
-	Mode            uint8  `json:"mode" binding:"required"`
+	Mode            int    `json:"mode" binding:"required"`
 }
